refactor: purgomalum infra package 추가
