+++ src/main/java/kitchenpos/products/infra/purgomalum/PurgomalumClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.products.infra;
+package kitchenpos.products.infra.purgomalum;
 
 public interface PurgomalumClient {
     boolean containsProfanity(String text);
